Decrease elastic rtq actualizer batch thresholds.
Now in case of lagging blades, actualizer takes too large batches to
process (10^7 events). Because it doesn't update offset until the end of batch,
it looses the whole work, which was already done, when it fails processing
somewhere in the middle of the batch (for example, because of unstable elastic)

Time threshold seems to be unreachable, but adjusted too just for consistency.